+++ R/markdataholo.R
@@ -1,4 +1,6 @@
-#' @title Mark data Holo.
+#' @title Mark Positional data - Holocen.
+#' @description When several OTUs, some can be monocen. and others holocen. Marks distance for monocen. are measured from cen. and for
+#' holocen. from top or bottom depending on \code{param} \code{origin}. See vignettes.
 #'
 #' @docType data
 #' @name markdataholo
@@ -10,13 +12,12 @@ NULL
 #'   \item{OTU}{OTU, species, optional}
 #'   \item{chrName}{name of chromosome}
 #'   \item{markName}{name of mark}
-#'   \item{markPos}{position from bottom or top (see parameter origin in plotIdiogramsHolo)}
+#'   \item{markPos}{position from bottom or top (see parameter \code{origin} in plotIdiograms)}
 #'   \item{markSize}{size of mark in micrometers or Mb}
-#'   \item{group}{chr. group, optional}
 #' }
-#' @seealso \code{\link{plotIdiogramsHolo}}
+#' @seealso \code{\link{markpos}}
+#' @seealso \code{\link{plotIdiograms}}
 #' @seealso \code{\link{chrbasicdataHolo}}
-#'
 #'
 #' @rdname markdataholo
 "bigdfMarkPosHolo"